remove int()

+++ corehq/apps/callcenter/checks.py
@@ -61,7 +61,7 @@ def _compile_stats(name, ucr_count, es_count):
     total = ucr_count + es_count
 
     def _percent(count):
-        return 0 if total == 0 else int(100 * count // total)
+        return 0 if total == 0 else 100 * count // total
 
     ucr_percent = _percent(ucr_count)
     es_percent = _percent(es_count)
